# - [*] move regexp to global vars to reduce allocs