This PR moves `regexp.MustCompile` to global variables to avoid compiling regular expressions every time `FormatXML` is called.

This results in a decreasing number of allocations per operation. Benchmarks show 60% fewer allocations.

```
$ git switch master
Already on 'master'
Your branch is up to date with 'upstream/master'.
$ go test -benchmem -bench=. -count=6 > old.txt
$ git switch refactor-compile-re-once
Switched to branch 'refactor-compile-re-once'
Your branch is up to date with 'origin/refactor-compile-re-once'.
$ go test -benchmem -bench=. -count=6 > new.txt
$ benchstat old.txt new.txt
```

<details>
<summary>Details</summary>

```
goos: darwin
goarch: arm64
pkg: github.com/go-xmlfmt/xmlfmt
cpu: Apple M1 Pro
                        │    old.txt     │               new.txt               │
                        │     sec/op     │    sec/op     vs base               │
FormatXML/XML1_1-8        10.501µ ±   5%   7.708µ ± 13%  -26.60% (p=0.002 n=6)
FormatXML/XML2_1-8         14.47µ ±   1%   11.46µ ±  7%  -20.76% (p=0.002 n=6)
FormatXML/XML3_1-8         23.90µ ±   1%   21.10µ ±  9%  -11.70% (p=0.002 n=6)
FormatXML/XML1_10-8       10.481µ ±  14%   7.799µ ±  4%  -25.59% (p=0.002 n=6)
FormatXML/XML2_10-8        14.71µ ±   8%   12.27µ ± 12%  -16.64% (p=0.002 n=6)
FormatXML/XML3_10-8        23.86µ ±  15%   22.69µ ±  8%        ~ (p=0.240 n=6)
FormatXML/XML1_100-8      10.389µ ±   1%   7.709µ ±  4%  -25.80% (p=0.002 n=6)
FormatXML/XML2_100-8       14.44µ ±   0%   11.62µ ± 28%        ~ (p=0.065 n=6)
FormatXML/XML3_100-8       23.83µ ±   0%   21.80µ ±  5%   -8.52% (p=0.002 n=6)
FormatXML/XML1_1000-8     10.440µ ±   1%   8.002µ ± 19%  -23.35% (p=0.002 n=6)
FormatXML/XML2_1000-8      15.62µ ± 120%   12.31µ ±  5%  -21.20% (p=0.002 n=6)
FormatXML/XML3_1000-8      24.84µ ± 130%   22.00µ ±  5%  -11.44% (p=0.002 n=6)
FormatXML/XML1_10000-8    10.738µ ±   6%   7.648µ ±  9%  -28.78% (p=0.002 n=6)
FormatXML/XML2_10000-8     14.65µ ±   5%   11.64µ ±  5%  -20.56% (p=0.002 n=6)
FormatXML/XML3_10000-8     23.90µ ±   6%   21.08µ ±  7%  -11.82% (p=0.002 n=6)
FormatXML/XML1_100000-8   10.626µ ±  11%   7.646µ ±  6%  -28.04% (p=0.002 n=6)
FormatXML/XML2_100000-8    14.79µ ±  14%   11.71µ ±  6%  -20.84% (p=0.002 n=6)
FormatXML/XML3_100000-8    27.31µ ±   7%   22.71µ ±  7%  -16.85% (p=0.002 n=6)
geomean                    15.64µ          12.61µ        -19.34%

                        │   old.txt    │               new.txt               │
                        │     B/op     │     B/op      vs base               │
FormatXML/XML1_1-8        6.995Ki ± 0%   2.430Ki ± 0%  -65.26% (p=0.002 n=6)
FormatXML/XML2_1-8        6.960Ki ± 0%   2.403Ki ± 0%  -65.47% (p=0.002 n=6)
FormatXML/XML3_1-8        7.993Ki ± 0%   3.436Ki ± 0%  -57.01% (p=0.002 n=6)
FormatXML/XML1_10-8       6.995Ki ± 0%   2.431Ki ± 0%  -65.25% (p=0.002 n=6)
FormatXML/XML2_10-8       6.963Ki ± 0%   2.403Ki ± 0%  -65.49% (p=0.002 n=6)
FormatXML/XML3_10-8       7.993Ki ± 0%   3.436Ki ± 0%  -57.02% (p=0.002 n=6)
FormatXML/XML1_100-8      6.993Ki ± 0%   2.428Ki ± 0%  -65.28% (p=0.002 n=6)
FormatXML/XML2_100-8      6.957Ki ± 0%   2.402Ki ± 0%  -65.47% (p=0.002 n=6)
FormatXML/XML3_100-8      7.993Ki ± 0%   3.434Ki ± 0%  -57.04% (p=0.002 n=6)
FormatXML/XML1_1000-8     6.996Ki ± 0%   2.430Ki ± 0%  -65.26% (p=0.002 n=6)
FormatXML/XML2_1000-8     6.962Ki ± 0%   2.402Ki ± 0%  -65.49% (p=0.002 n=6)
FormatXML/XML3_1000-8     7.988Ki ± 0%   3.436Ki ± 0%  -56.99% (p=0.002 n=6)
FormatXML/XML1_10000-8    6.997Ki ± 0%   2.428Ki ± 0%  -65.29% (p=0.002 n=6)
FormatXML/XML2_10000-8    6.965Ki ± 0%   2.401Ki ± 0%  -65.52% (p=0.002 n=6)
FormatXML/XML3_10000-8    7.991Ki ± 0%   3.436Ki ± 0%  -57.00% (p=0.002 n=6)
FormatXML/XML1_100000-8   6.992Ki ± 0%   2.430Ki ± 0%  -65.25% (p=0.002 n=6)
FormatXML/XML2_100000-8   6.963Ki ± 0%   2.402Ki ± 0%  -65.50% (p=0.002 n=6)
FormatXML/XML3_100000-8   7.992Ki ± 0%   3.436Ki ± 0%  -57.00% (p=0.002 n=6)
geomean                   7.301Ki        2.717Ki       -62.79%

                        │  old.txt   │              new.txt              │
                        │ allocs/op  │ allocs/op   vs base               │
FormatXML/XML1_1-8        79.00 ± 0%   29.00 ± 0%  -63.29% (p=0.002 n=6)
FormatXML/XML2_1-8        70.00 ± 0%   20.00 ± 0%  -71.43% (p=0.002 n=6)
FormatXML/XML3_1-8        69.00 ± 0%   19.00 ± 0%  -72.46% (p=0.002 n=6)
FormatXML/XML1_10-8       79.00 ± 0%   29.00 ± 0%  -63.29% (p=0.002 n=6)
FormatXML/XML2_10-8       70.00 ± 0%   20.00 ± 0%  -71.43% (p=0.002 n=6)
FormatXML/XML3_10-8       69.00 ± 0%   19.00 ± 0%  -72.46% (p=0.002 n=6)
FormatXML/XML1_100-8      79.00 ± 0%   29.00 ± 0%  -63.29% (p=0.002 n=6)
FormatXML/XML2_100-8      70.00 ± 0%   20.00 ± 0%  -71.43% (p=0.002 n=6)
FormatXML/XML3_100-8      69.00 ± 0%   19.00 ± 0%  -72.46% (p=0.002 n=6)
FormatXML/XML1_1000-8     79.00 ± 0%   29.00 ± 0%  -63.29% (p=0.002 n=6)
FormatXML/XML2_1000-8     70.00 ± 0%   20.00 ± 0%  -71.43% (p=0.002 n=6)
FormatXML/XML3_1000-8     69.00 ± 0%   19.00 ± 0%  -72.46% (p=0.002 n=6)
FormatXML/XML1_10000-8    79.00 ± 0%   29.00 ± 0%  -63.29% (p=0.002 n=6)
FormatXML/XML2_10000-8    70.00 ± 0%   20.00 ± 0%  -71.43% (p=0.002 n=6)
FormatXML/XML3_10000-8    69.00 ± 0%   19.00 ± 0%  -72.46% (p=0.002 n=6)
FormatXML/XML1_100000-8   79.00 ± 0%   29.00 ± 0%  -63.29% (p=0.002 n=6)
FormatXML/XML2_100000-8   70.00 ± 0%   20.00 ± 0%  -71.43% (p=0.002 n=6)
FormatXML/XML3_100000-8   69.00 ± 0%   19.00 ± 0%  -72.46% (p=0.002 n=6)
geomean                   72.53        22.25       -69.32%
```

</p>
</details>